refactor: relocate navigation controller implementations to feature modules
Move default navigation controller implementations from the centralized `:navigation:implementation` module to their respective feature modules to improve encapsulation and modularity.

- **Relocation**:
    - Move `DefaultEpisodeSheetNavigator` to `:features:episode-sheet:presenter`.
    - Move `DefaultHomeTabNavigator` to `:features:home:presenter`.
    - Move `DefaultNotificationRationale` and `DefaultSheetNavigator` to `:features:root:presenter`.
- **Module Cleanup**: Remove feature-specific navigation API dependencies from the `commonMain` source set of `:navigation:implementation`, shifting necessary dependencies to `commonTest`.
- **Testing & DI**: Update `FakeAppBindings` and project imports to reflect the new package structures for the relocated controllers.
