+++ features/root/presenter/src/commonMain/kotlin/com/thomaskioko/tvmaniac/presenter/root/di/DefaultNotificationRationale.kt
@@ -1,4 +1,4 @@
-package com.thomaskioko.tvmaniac.navigation.controllers
+package com.thomaskioko.tvmaniac.presenter.root.di
 
 import com.thomaskioko.root.nav.NotificationRationale
 import com.thomaskioko.tvmaniac.core.base.ActivityScope
