+++ navigation/implementation/build.gradle.kts
@@ -12,15 +12,8 @@ scaffold {
 kotlin {
     sourceSets {
         commonMain.dependencies {
-            implementation(projects.features.root.nav)
             implementation(projects.core.base)
             implementation(projects.navigation.api)
-            implementation(projects.data.datastore.api)
-
-            implementation(projects.features.episodeSheet.nav)
-            implementation(projects.features.home.nav)
-            implementation(projects.features.showDetails.nav)
-            implementation(projects.features.seasonDetails.nav)
 
             implementation(libs.coroutines.core)
         }
@@ -29,7 +22,11 @@ kotlin {
             implementation(projects.core.util.testing)
             implementation(projects.data.traktauth.testing)
             implementation(projects.core.testing.di)
+            implementation(projects.features.episodeSheet.nav)
             implementation(projects.features.genreShows.nav)
+            implementation(projects.features.home.nav)
+            implementation(projects.features.seasonDetails.nav)
+            implementation(projects.features.showDetails.nav)
 
             implementation(libs.bundles.unittest)
         }
